+++ lib/screen_layout.dart
@@ -1,6 +1,16 @@
 library screen_layout;
 
-export 'src/device_utils.dart';
-export 'src/enums.dart';
-export 'src/app_screen_state.dart';
-export 'src/app_widget_state.dart';
+import 'package:flutter/foundation.dart' show defaultTargetPlatform, kIsWeb;
+import 'package:flutter/widgets.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter/foundation.dart';
+
+part 'src/device_utils.dart';
+part 'src/enums.dart';
+part 'src/app_screen_state.dart';
+part 'src/app_widget_state.dart';
+part 'src/base_screen.dart';
+
+part 'src/extension.dart';
+part 'src/helper.dart';
+part 'src/widget.dart';