+++ lib/src/app_widget_state.dart
@@ -1,7 +1,4 @@
-import 'package:flutter/material.dart';
-import 'package:screen_layout/src/base_screen.dart';
-import 'package:screen_layout/src/device_utils.dart';
-import 'package:screen_layout/src/enums.dart';
+part of screen_layout;
 
 /// most basic app screen state handles the screen orientations
 /// and user can build portrait design in [buildPortraitLayout]
@@ -13,22 +10,18 @@ abstract class AppWidgetState extends StatelessWidget
 
   @override
   Widget build(BuildContext context) {
-    return LayoutBuilder(builder: (context, boxConstraints) {
-      MediaQueryData mediaQuery = MediaQuery.of(context);
-      DeviceScreenType _deviceType = getDeviceType(mediaQuery);
-      Orientation _orientation = mediaQuery.orientation;
-      Size _screenSize = mediaQuery.size;
-
+    return LayoutBuilder(builder: (_context, boxConstraints) {
       DeviceUtils _deviceUtils = DeviceUtils(
-          screenType: _deviceType,
-          screenSize: _screenSize,
-          orientation: _orientation);
+          deviceType: Device.deviceType,
+          screenType: Device.screenType,
+          screenSize: Size(boxConstraints.maxWidth, boxConstraints.maxHeight),
+          orientation: Device.orientation);
 
-      if (_deviceType == DeviceScreenType.desktop) {
+      if (_deviceUtils.screenType == ScreenType.desktop) {
         return MediaQuery.of(context).orientation == Orientation.portrait
             ? buildDesktopPortraitLayout(context, _deviceUtils)
             : buildDesktopLandscapeLayout(context, _deviceUtils);
-      } else if (_deviceType == DeviceScreenType.tablet) {
+      } else if (_deviceUtils.screenType == ScreenType.tablet) {
         return MediaQuery.of(context).orientation == Orientation.portrait
             ? buildTabletPortraitLayout(context, _deviceUtils)
             : buildTabletLandscapeLayout(context, _deviceUtils);