+++ lib/src/enums.dart
@@ -1,1 +1,9 @@
-enum DeviceScreenType { mobile, tablet, desktop, watch }
+part of screen_layout;
+
+/// Type of Device
+/// This can be mobile tablet, desktop or watch
+enum ScreenType { mobile, tablet, desktop, watch }
+
+/// Type of Operating System
+/// This can be android, ios, fuschia, web, or desktop (windows, mac, linux)
+enum DeviceType { android, ios, fuschia, web, windows, mac, linux }