+++ CHANGELOG.md
@@ -5,3 +5,11 @@ Added AppScreenState and AppWidgetState abstract class for making responsive app
 ## 0.0.2
 
 Added deviceType support in AppScreenState and Example project in github
+
+## 0.0.3
+
+Added types for screensize, orientation and deviceType
+
+## 0.0.4
+
+Added ResponsiveLayout Builder, DeviceType,  ScreenType, and DeviceExt